Update impact and progress values to use ushort type
The update changes the data type of impact and progress fields in the TaskProgressService and TaskProgressItem classes from double to ushort. Other adjustments were made in related methods and classes to ensure compatibility with this change. Moreover, the indeterminate state condition of the progress bar was also updated. This change simplifies the code structure and enhances the readability and maintainability of the system.

+++ Spravy.Ui/Extensions/TaskProgressServicePropertyExtension.cs
@@ -4,7 +4,7 @@ public static class TaskProgressServicePropertyExtension
 {
     public static ConfiguredValueTaskAwaitable<Result> RunProgressAsync<TTaskProgressServiceProperty>(
         this TTaskProgressServiceProperty property,
-        double impact,
+        ushort impact,
         Func<TaskProgressItem, ConfiguredValueTaskAwaitable<Result>> func,
         CancellationToken cancellationToken
     ) where TTaskProgressServiceProperty : ITaskProgressServiceProperty
